fix(core): various auth fixes
