+++ examples/blog/.env.example
@@ -1,11 +1,13 @@
-# MongoDB
+### MongoDB
 MONGODB_URI=mongodb://localhost/catalyst-blog-example
 
-# Next Auth
-NEXTAUTH_SECRET="SOME-SECRET-HERE" # this can be any random unguessable string
-NEXTAUTH_URL="http://localhost:3000" # points to the nextjs instance i.e. localhost in dev and actual domain in production
+### Next Auth
+# this can be any random unguessable string
+NEXTAUTH_SECRET="SOME-SECRET-HERE"
+# points to the nextjs instance i.e. localhost in dev and actual domain in production
+NEXTAUTH_URL="http://localhost:3000"
 
-# Google OAuth
+### Google OAuth
 # Note: hit me up if you want to use my google oauth credentials for dev
 GOOGLE_CLIENT_ID="some-google-oauth-id"
 GOOGLE_CLIENT_SECRET="some-google-oauth-secret"
